backend/requirements.txt
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pandas>=2.2.0
openpyxl>=3.1.0
xlrd==1.2.0
openai>=1.0.0
